requirements.txt
playwright==1.42.0
python-telegram-bot==20.8
schedule==1.2.1
python-dotenv==1.0.1
